+++ CHANGELOG.md
@@ -11,6 +11,12 @@ All notable changes to this project will be documented in this file with the fol
 ### Fixed
 - Bug fixes.
 
+## [3.3.0] - 2026-04-27
+
+### Added
+- The `terralab deliver <JOB ID> <GCS_DESTINATION>` command allows users to deliver pipeline outputs to a specified Google Cloud Storage location.
+- The `terralab jobs details <JOB ID>` command has been updated to display information about data delivery.
+
 ## [3.2.0] - 2026-04-13
 
 ### Added